Move rotate tests into /pkg/rotate
